neurons/text_to_video/miners/requirements.txt
modelscope==1.4.2
open_clip_torch
pytorch-lightning